moved import to /test (it's the same for callback&promise)
